requirements.txt
ansicolors==1.1.8
docker_py==1.10.6
